create config/schemas/ and move well-scoped schema files into it
