+++ assistant/src/config/schemas/elevenlabs.ts
