+++ assistant/src/config/schemas/mcp.ts
