+++ assistant/src/config/schemas/skills.ts
